EmployeeManagementSystem/employees/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import EmployeeViewSet, DepartmentViewSet, RoleViewSet, login_view


router = DefaultRouter()
router.register(r'employees', EmployeeViewSet, basename='employee')
router.register(r'departments', DepartmentViewSet, basename='department')
router.register(r'roles', RoleViewSet, basename='role')


urlpatterns = [
    path('auth/login/', login_view, name='login'),
]

urlpatterns += router.urls

