requirements.txt
Django==6.0
djangorestframework==3.15.2
django-filter==24.3
PyJWT==2.9.0
django-redis==5.4.0
django-debug-toolbar==4.4.6
coverage==7.6.1
mysqlclient==2.2.4